Fix Z-buffer scroll-clear stride + COLORFILL truncation (audit K2,V2)
- renderworld ScrollBackground: the !gfDoVideoScroll branch pre-cleared the
  Z-buffer with SCREEN_WIDTH*2 (the 16bpp stride), under-clearing the lower
  viewport on 32bpp and leaving stale depth. RenderStaticWorldRect (called on
  the next line) already clears it at the correct uiDestPitchBYTES, so the wrong
  pre-clear is removed as both incorrect and redundant.
- sdl_vsurface BltVideoSurfaceToVideoSurface (VS_BLT_COLORFILL): dropped the
  (UINT16) casts that truncated the 32-bit ColorFill before PixFromColor16
  (would have re-decoded the survivors as RGB565 -> wrong/faded fill).

+++ sgp/sdl_vsurface.cpp
@@ -865,7 +865,7 @@ BOOLEAN BltVideoSurfaceToVideoSurface(HVSURFACE hDst, HVSURFACE hSrc,
 			PRIMARY_SURFACE, // unused -- we go through GetVideoSurface dispatch below
 			pBltFx->FillRect.iLeft, pBltFx->FillRect.iTop,
 			pBltFx->FillRect.iRight, pBltFx->FillRect.iBottom,
-			(UINT16)pBltFx->ColorFill) ? TRUE : FALSE;
+			pBltFx->ColorFill) ? TRUE : FALSE;
 	}
 	if (fBltFlags & VS_BLT_COLORFILL)
 	{
@@ -874,7 +874,7 @@ BOOLEAN BltVideoSurfaceToVideoSurface(HVSURFACE hDst, HVSURFACE hSrc,
 		PIXEL* buf = (PIXEL*)LockVideoSurfaceBuffer(hDst, &pitch);
 		if (!buf) return FALSE;
 		FillRect16(buf, pitch, 0, 0, hDst->usWidth, hDst->usHeight,
-		           (UINT16)pBltFx->ColorFill);
+		           pBltFx->ColorFill);
 		UnLockVideoSurfaceBuffer(hDst);
 		return TRUE;
 	}
